Started refactoring the code since its too big to maintain
